+++ e2e/config.js
@@ -16,6 +16,9 @@ module.exports = {
     ],
     run: [
         {
+            groupName: 'notifications'
+        },
+        {
             groupName: 'system'
         },
         {