+++ packages/web-platform/src/shared/decoders.ts
@@ -71,13 +71,14 @@ export const windowLayoutComponentDecoder: Decoder<Glue42Web.Layouts.WindowCompo
     })
 });
 
-export const libDomainDecoder: Decoder<LibDomains> = oneOf<"system" | "windows" | "appManager" | "layouts" | "workspaces" | "intents">(
+export const libDomainDecoder: Decoder<LibDomains> = oneOf<"system" | "windows" | "appManager" | "layouts" | "workspaces" | "intents" | "notifications">(
     constant("system"),
     constant("windows"),
     constant("appManager"),
     constant("layouts"),
     constant("workspaces"),
-    constant("intents")
+    constant("intents"),
+    constant("notifications")
 );
 
 export const systemOperationTypesDecoder: Decoder<SystemOperationTypes> = oneOf<"getEnvironment" | "getBase">(
@@ -305,12 +306,18 @@ export const windowsConfigDecoder: Decoder<Glue42WebPlatform.Windows.Config> = o
     }))
 });
 
+export const serviceWorkerConfigDecoder: Decoder<Glue42WebPlatform.ServiceWorker.Config> = object({
+    url: optional(nonEmptyStringDecoder),
+    registrationPromise: optional(anyJson())
+});
+
 export const platformConfigDecoder: Decoder<Glue42WebPlatform.Config> = object({
     windows: optional(windowsConfigDecoder),
     applications: optional(applicationsConfigDecoder),
     layouts: optional(layoutsConfigDecoder),
     channels: optional(channelsConfigDecoder),
     plugins: optional(pluginsConfigDecoder),
+    serviceWorker: optional(serviceWorkerConfigDecoder),
     gateway: optional(gatewayConfigDecoder),
     glue: optional(glueConfigDecoder),
     workspaces: optional(workspacesConfigDecoder),