+++ packages/web-worker/src/index.ts
@@ -1,0 +1,10 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { setupCore, openCorePlatform, raiseGlueNotification } from "./worker";
+
+if (typeof self !== "undefined") {
+    (self as any).GlueWebWorker = setupCore;
+    (self as any).openCorePlatform = openCorePlatform;
+    (self as any).raiseGlueNotification = raiseGlueNotification;
+}
+
+export default setupCore;