+++ packages/web-platform/changelog.md
@@ -1,4 +1,6 @@
-1.5.1
+1.6.0
+feat: added support for a Service Worker
+feat: added support for advanced notifications
 fix: fixed allowDropLeft, allowDropTop, allowDropRight, allowDropBottom in the workspaces config object by adding them to the decoders
 1.5.0
 feat: extended workspaces protocol and decoders for servicing frame bounds requests